# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AutoLlama v2.2 - Intelligent Context API
  description: |
    **AutoLlama v2.2 "Intelligent Context Llama"** - Enhanced RAG platform with intelligent contextual retrieval and advanced document understanding.
    
    ## 🚀 Enhanced Features v2.2
    - **35-60% Better Retrieval** through intelligent contextual processing
    - **Document Intelligence**: Automatic type detection and structural analysis
    - **Semantic Chunking**: Boundary-aware segmentation with hierarchical understanding
    - **Advanced Context Generation**: Document-aware prompts with retry logic
    - **Enhanced Metadata**: 11 new contextual fields for performance analysis
    - **Production-Ready**: Ultra-safe processing with comprehensive error handling
    
    ## Processing Capabilities
    - **Multi-format Processing**: PDF, DOCX, EPUB, CSV, URL fetching
    - **Hybrid Search**: Vector similarity + BM25 text search  
    - **Real-time Processing**: SSE streaming with progress updates
    - **Intelligent Chunking**: Semantic, structural, and hierarchical strategies
    - **Contextual Analysis**: Sentiment, topics, emotions, categories + document structure
    
    ## Authentication
    This API is designed for secure homelab deployment behind Tailscale VPN.
    
    ## Rate Limiting & Error Handling
    - **Search endpoints**: 100 requests/minute
    - **Processing endpoints**: 10 requests/5 minutes  
    - **Health endpoints**: No limits
...
  version: 2.2.0
...
  contact:
...
servers:
...
paths:
...
  /process-url:
    post:
      tags: [Content Processing]
      summary: Process URL synchronously
      description: |
        📄 Process content from a URL and return when complete.
        
        **Supported Content:**
        - Articles, blog posts, documentation
        - PDF documents, research papers  
        - Wikipedia pages, news articles
        
        **Try these example URLs:**
        - https://docs.python.org/3/tutorial/
        - https://en.wikipedia.org/wiki/Machine_learning
        - Any public webpage or PDF URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/URLProcessRequest'
            example:
              url: "https://docs.python.org/3/tutorial/introduction.html"
              chunkSize: 1000
              overlap: 100
              enableContextualEmbeddings: true
      responses:
        '200':
          description: URL processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ProcessingError'
  /process-url-stream:
    post:
      tags: [Content Processing]
      summary: Process URL with streaming
      description: |
        Process URL content with real-time progress updates via Server-Sent Events.
        Returns immediately with session ID for progress tracking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/URLProcessRequest'
      responses:
        '200':
          description: Processing started, use SSE for progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamingResponse'
  /process-file:
    post:
      tags: [Content Processing]
      summary: Process file synchronously
      description: |
        Upload and process a file. Supports PDF, DOCX, EPUB, CSV, and text files.
        Maximum file size: 100MB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileProcessRequest'
            encoding:
              file:
                contentType: application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/epub+zip, text/csv, text/plain
      responses:
        '200':
          description: File processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '413':
          description: File size exceeds limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /process-file-stream:
    post:
      tags: [Content Processing]
      summary: Process file with streaming
      description: Process file with real-time progress updates via SSE
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileProcessRequest'
      responses:
        '200':
          description: Processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamingResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "⚠️ NOT IMPLEMENTED - Leave empty for local development"
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "⚠️ NOT IMPLEMENTED - Leave empty for local development"
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Validation failed"
              code: "VALIDATION_ERROR"
              details:
                - field: "chunkSize"
                  message: "Must be between 100 and 5000"
            statusCode: 400
...
    ProcessingError:
      description: Processing failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Failed to process content"
              code: "PROCESSING_ERROR"
              details: "Unable to parse PDF content"
            statusCode: 500
  schemas:
...
    URLProcessRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: "https://docs.python.org/3/tutorial/introduction.html"
          description: "🔗 Any public URL - Try Wikipedia, documentation sites, or blog posts"
        chunkSize:
          type: integer
          minimum: 100
          maximum: 5000
          default: 1000
          description: "📝 Text chunk size (1000 = ~2-3 paragraphs, good default)"
        overlap:
          type: integer
          minimum: 0
          maximum: 500
          default: 100
          description: "🔄 Character overlap between chunks (100 = ~1 sentence, prevents context loss)"
        enableContextualEmbeddings:
          type: boolean
          default: true
          description: "🧠 Generate smarter embeddings with context (recommended: true)"
    FileProcessRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: File to process (PDF, DOCX, EPUB, CSV, TXT)
        chunkSize:
          type: integer
          minimum: 100
          maximum: 5000
          default: 1000
        overlap:
          type: integer
          minimum: 0
          maximum: 500
          default: 100
        enableContextualEmbeddings:
          type: boolean
          default: true
...
    ProcessingResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
          format: uuid
        message:
          type: string
        results:
          type: object
          properties:
            documentsCreated:
              type: integer
            chunksProcessed:
              type: integer
            totalTokensUsed:
              type: integer
            processingTime:
              type: number
    StreamingResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
          format: uuid
        message:
          type: string
        sseEndpoint:
          type: string
          description: Server-Sent Events endpoint for progress updates
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              nullable: true
            details:
              oneOf:
                - type: string
                - type: object
                - type: array
              nullable: true
        statusCode:
          type: integer

# No global security required - this is a local development API